--- src/app/api/config/config.json
@@ -1,5 +1,0 @@
-{
-  "flags": {
-    "showUserInfoInHeader": true
-  }
-}
