--- src/app/api/config/route.ts
@@ -1,15 +1,0 @@
-import Config from "./config.json";
-
-/**
- * @name api/config endpoint
- * @async
- * @api
- *
- * @returns Config Data
- *
- */
-export async function GET() {
-  return Response.json(Config, { status: 200 });
-}
-
-// This should fetch from Server in the Future
