+++ src/app/config.json
@@ -1,0 +1,9 @@
+{
+  "flags": {
+    "showUserInfoInHeader": true
+  },
+  "gates": {
+    "nodata": null
+  },
+  "wip": true
+}
